Undo moving utils function
