+++ sdv/utils.py
@@ -184,3 +184,12 @@ def load_data_from_csv(filepath, pandas_kwargs=None):
 def groupby_list(list_to_check):
     """Return the first element of the list if the length is 1 else the entire list."""
     return list_to_check[0] if len(list_to_check) == 1 else list_to_check
+
+
+def create_unique_name(name, list_names):
+    """Modify the ``name`` parameter if it already exists in the list of names."""
+    result = name
+    while result in list_names:
+        result += '_'
+
+    return result
